+++ tests/Feature/RadioTest.php
@@ -13,6 +13,16 @@ class RadioTest extends TestCase
         $this->registerTestRoute('default-radio');
 
         $this->visit('/default-radio')
+            ->seeElement('input[value="1"]:checked')
+            ->seeElement('input[value="0"]:not(:checked)');
+    }
+
+    /** @test */
+    public function it_check_the_right_element_as_default_with_a_bound_target()
+    {
+        $this->registerTestRoute('default-radio-bind');
+
+        $this->visit('/default-radio-bind')
             ->seeElement('input[value="a"]:checked')
             ->seeElement('input[value="b"]:not(:checked)');
     }