+++ tests/Feature/views/default-radio-bind.blade.php
@@ -3,13 +3,15 @@ $target = ['radio' => 'a'];
 @endphp
 
 <x-form>
+    @bind($target)
     <x-form-group>
-        <x-form-radio name="radio" value="1" default />
+        <x-form-radio name="radio" value="a" />
     </x-form-group>
 
     <x-form-group>
-        <x-form-radio name="radio" value="0" />
+        <x-form-radio name="radio" value="b" />
     </x-form-group>
+    @endbind
 
     <x-form-submit />
 </x-form>